v0.1.5
BINTRICALIZATION

+++ gameProgramming/gaming_exercises/Cohort_Lessons/07_DNAGame/dnagame.py
@@ -77,7 +77,7 @@ def calcScore(rnaSequence: str, rnaTime: float) -> float:
     score *= scoreMulti
     return score
 
-def saveScore(dnaSequence: str, rnaSequence: str, rnaTime: float, score: float) -> None:
+def saveScore(dnaSequence: str, rnaSequence: str, rnaTime: int, score: float) -> None:
     playerName = input("What is your first name?\n")
     lastName = input("What is your last name?\n")
     fullName = playerName + "  " + lastName
